majority of styles, layout adjustments in game section & past responses, added scroll bar to Description section
