+++ package.json
@@ -14,6 +14,7 @@
     "firebase": "^9.16.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.7.1",
     "react-scripts": "5.0.1",
     "sass": "^1.58.0",
     "web-vitals": "^2.1.4"
