+++ src/Components/Footer.js
@@ -1,0 +1,14 @@
+const Footer = () => {
+
+  return (
+    <footer>
+      <div className="footer-container wrapper">
+        <p>Created by Fiona Freemark for 
+          <a href="https://www.junocollege.com"> Juno College of Technology</a>
+        </p>
+      </div>
+    </footer>
+  )
+}
+
+export default Footer;
