--- src/sass/partials/_cardDisplay.scss
@@ -1,1 +1,0 @@
-@use "./variables" as *;
