+++ src/sass/partials/_gameSection.scss
@@ -1,0 +1,202 @@
+@use "./variables" as *;
+
+.game-container-icon {
+  position: absolute;
+  bottom: 5%;
+  left: 48%;
+}
+
+.game-container {
+  min-height: 980px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: $light-blue;
+  position: relative;
+}
+
+.game-content-box {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  flex-wrap: wrap;
+  min-height: 70%;
+  width: 100%;
+  padding: 15% 0 20% 0;
+  position: relative;
+}
+
+.game-left-box {
+  flex-basis: 25%;
+  height: 100%;
+}
+
+.card-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  .button {
+    align-self: center;
+    margin-top: 5%;
+  }
+}
+
+.tarot-card {
+  border: 15px white solid;
+  margin: 8px;
+  border-radius: 10px;
+  transition: 300ms;
+}
+
+// .tarot-card-flip-to {
+//   animation: flip-to 2s ease-in-out;
+// }
+
+// .tarot-card-flip {
+// animation: flip-from 2s ease-in-out;
+// }
+
+// @keyframes flip-to {
+
+//   to {
+//     transform: rotateY(180deg);
+//   }
+// }
+// @keyframes flip-from {
+
+//   from {
+//     transform: rotateY(-180deg);
+//   }
+// }
+
+// RIGHT BOX
+.game-right-outer-box {
+  flex-basis: 60%;
+  height: 100%;
+}
+
+.game-right-box {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+  height: 100%;
+}
+
+.card-info-box {
+      padding: 2% 0;
+}
+
+.drawn-card-name {
+  margin-bottom: 3%;
+}
+
+.results-header,
+.results-text {
+  padding: 2%;
+}
+
+.results-text {
+  max-height: 200px;
+  text-overflow: ellipsis;
+  overflow: scroll;
+}
+
+// SAVE YOUR CARD SECTION
+
+form {
+    margin-top: 2%;
+    padding-bottom: 10%;
+}
+
+.initials-box {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.initials-label {
+  padding: 2%;
+  font-weight: 400;
+  text-transform: uppercase;
+  font-weight: 600;
+}
+
+.initials-input {
+  padding: 1%;
+  border-radius: 2px;
+  background-color: $transparent;
+  border: 1px solid $black;
+  width: 15%;
+  text-align: center;
+}
+
+legend {
+  text-align: center;
+  display: inline-block;
+  text-transform: uppercase;
+  width: 100%;
+  margin-top: 5%;
+  font-weight: 600;
+}
+
+
+fieldset {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 2%
+}
+
+fieldset label {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-basis: 5%;
+  margin: 0 2%;
+}
+
+input[type="radio"] {
+  opacity: 0;
+  position: absolute;
+  z-index: -1;
+}
+.icon {
+  padding: 10px;
+  font-size: 3rem;
+  position: relative;
+  border: 1px solid $transparent;
+  border-radius: 2px;
+  transition: 300ms;
+}
+
+label:hover,
+label:focus,
+label:active
+input[type="radio"]:checked+label,
+input[type="radio"]:focus+label,
+input[type="radio"]:active+label {
+color: $black;
+background-color: $yellow;
+border: 1px solid $black;
+border-radius: 2px;
+}
+
+.submit-button-box {
+  text-align: center;
+}
+
+// Arrow pointing downward at bottom:
+.past-response-arrow-box {
+display: flex;
+flex-direction: column;
+align-items: center;
+justify-content: center;
+position: absolute;
+bottom: 2%;
+left: 42%
+}
