+++ src/sass/partials/_pastResults.scss
@@ -1,34 +1,68 @@
 @use "./variables" as *;
 
+.past-responses {
+  background-color: $neutral;
+  padding: 10% 0;
+}
 
+.past-response-heading {
+  margin: 2% 0;
+}
 
 .log-ul {
   display: flex;
   justify-content: space-between;
-  /* align-items: flex-start; */
   flex-wrap: wrap;
   list-style: none;
 }
 
 .log-li {
   flex-basis: 28%;
+  margin: 5% 1%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.saved-img-container {
   position: relative;
-  margin: 5% 0;
 }
 
 .log-initials {
   text-transform: uppercase;
   font-weight: 600;
+  padding-left: 5px;
 }
 
-.log-icon {
-  background-color: white;
-  border-radius: 50%;
-  position: absolute;
-  left: 25%;
-  bottom: -5%;
+.emoji-log-box {
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
 }
 
 .tarot-log-img {
   max-width: 150px;
+  border: 8px solid white;
+  border-radius: 10px;
+}
+
+.button-container {
+  display: flex;
+  justify-content: center;
+}
+
+.submit-button {
+    background-color: $light-blue;
+      color: $black;
+      border: 1px solid $black;
+
+  margin: 10px;
+}
+
+.submit-button:hover,
+.submit-button:focus {
+  background-color: $black;
+    color: $white;
+    border: 1px solid $black;
 }
