+++ src/sass/partials/_variables.scss
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600&family=Yeseva+One&display=swap');
+
 // Colours
 $white: #fff;
 $black: #161312;
@@ -26,7 +28,5 @@ $fw-bold: 700;
 
 // Font Family
 $ff-headings: 'Yeseva One', cursive;
-/* 700 Italic */
 $ff-body: 'Montserrat',
 sans-serif;
-/* in $fw-default/ 600 */
